% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_heatmap.R
\name{annot_heatmap}
\alias{annot_heatmap}
\title{Annotated heatmap}
\usage{
annot_heatmap(mat, tidydf, annot1, annot2, title, scale, fontsize = 8)
}
\arguments{
\item{mat}{Numeric matrix where each row is an antibody and each column is
a sample. Column names should be sample names.}

\item{tidydf}{Tidy dataframe of RFI and annotation data.}

\item{annot1}{Name of the column in tidydf containing the annotation data,
as string.}

\item{annot2}{Optional argument. Name of the column in tidydf containing 
the second annotation data, as string.}

\item{title}{Title of the graph as string.}

\item{scale}{Logical. If TRUE, the heatmap will be scaled across rows 
(antibodies), if FALSE no scaling will be done.}

\item{fontsize}{Single numeric indicating the size of the font in the 
heatmap. Passed to \code{pheatmap()}.}
}
\description{
Generate an annotated heatmap using pheatmap packaged.
}
