% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldChange.R
\name{calcFC}
\alias{calcFC}
\alias{plotFC}
\title{Calculate fold change}
\usage{
calcFC(tidydf, comparisons, log = FALSE, RFI_col = "RFI", ABnames)

plotFC(fc_df, comparisons, logged = FALSE, normalised = FALSE)
}
\arguments{
\item{tidydf}{Tidy dataframe of RFI values and their corresponding sample 
and phenotype information. There should NOT be any technical replicates 
(samples with the same sample name).}

\item{comparisons}{Dataframe with 2 columns and n rows, where n is the 
number of desired comparisons. The fold change to be calculated will be
column 2 divided by column 1, for each comparison/row.}

\item{log}{Single logical indicating whether the fold change should be 
logged (base 2).}

\item{RFI_col}{Name of column containing RFI values, as string.}

\item{ABnames}{Optional argument. Merges full antibody names to the output
dataframe.}

\item{fc_df}{Dataframe of fold changes generated by \code{calcFC}.}

\item{logged}{Single logical indicating whether the data have been logged}

\item{normalised}{Single logical indicating whether the data have been 
normalised.}
}
\description{
Calculates the fold changes between samples for a given set of comparisons.
\code{plotFC} plots these fold changes as a bar graph.
}
\section{Functions}{
\itemize{
\item \code{calcFC}: Generates a dataframe of fold changes for each 
comparison, with columns of corresponding names of samples compared and 
their phenotype description.

\item \code{plotFC}: Generates a barplot of fold changes for the desired
comparisons.
}}

