% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_mat.R
\name{df_to_mat}
\alias{df_to_mat}
\title{Convert tidy dataframe into matrix}
\usage{
df_to_mat(tidydf, logdata, tech_reps, convert_neg = TRUE,
  AB_col = "Antibody.Name")
}
\arguments{
\item{tidydf}{Tidy dataframe of RFI values to convert to matrix.}

\item{logdata}{Single logical indicating whether the raw RFI values should
be logged.}

\item{tech_reps}{Single logical indicating whether there are technical
replicates.}

\item{convert_neg}{Single logical indicating whether negative values should
be converted to zero (0).}

\item{AB_col}{Name of the column containing antibody name or number,
as string. These names or numbers will become rownames of the 
output matrix.}
}
\value{
A numeric matrix where each row is an antibody and each column is
    a sample.
}
\description{
Converts a tidy dataframe into a numeric matrix.
}
\details{
If the tidy dataframe contains technical replicates with the same sample
names, unique sample names will be created by add a group row number to the
the end of the sample name.

Useful for functions that require data to be in matrix format like 
\code{pheatmap} and \code{plotMA}.
}
