% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRIN.R
\name{.GraphSetup}
\alias{.GraphSetup}
\title{Global rank invariant normalisation}
\usage{
.GraphSetup(width = 6, height = 6, pointsize = 12,
  ReportName = "test", filetype = "none")
}
\arguments{
\item{width}{Width (inches) of diagnostic plot.}

\item{height}{Height (inches) of diagnostic plot.}

\item{pointsize}{Point size for text of diagnostic plot.}

\item{ReportName}{Name for Diagnostic plots.}

\item{filetype}{Diagnostic plot format "png", "wmf", or "postscript"}

\item{data}{Matrix of RFI values with column for each sample.}

\item{count}{Size of Global Rank-invariant Set to use.}

\item{f}{Smoother parameter for lowess.}
}
\description{
Performs global rank invariant normalisation, as described by \href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-9-520}{Pelz et al.}
}
\details{
The code for this function was taken from the \href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-9-520}{Pelz et al.}. 
paper. Some minor changes have been made to allow it to be used for RPPA 
data.
}
