% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_GRIN.R
\name{GRSN}
\alias{GRSN}
\title{Global rank invariant normalisation (GRIN)}
\usage{
GRSN(data, width = 15, height = 5, pointsize = 16,
  filetype = "png", ReportName = NA, count = 5000, f = 0.25)
}
\arguments{
\item{data}{Matrix of RFI values with column for each sample.}

\item{width}{Width (inches) of diagnostic plot.}

\item{height}{Height (inches) of diagnostic plot.}

\item{pointsize}{Point size for text of diagnostic plot.}

\item{filetype}{Diagnostic plot format "png", "wmf", or "postscript"}

\item{ReportName}{Name for Diagnostic plots.}

\item{count}{Size of Global Rank-invariant Set to use.}

\item{f}{Smoother parameter for lowess.}
}
\description{
Performs global rank invariant normalisation, as described by \href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-9-520}{Pelz et al.}.
}
\details{
The code for this function was taken from \href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-9-520}{Pelz et al.} 
Some minor changes have been made to allow it to be used for RPPA 
data.
}
