% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputPheno.R
\name{inputPheno}
\alias{inputPheno}
\title{Create phenotype dataframe}
\usage{
inputPheno(path, sheet, col_remove, condition_cols)
}
\arguments{
\item{path}{Single string containing the path to the phenotyle xlsx file.}

\item{sheet}{Single numeric indicating the sheet number or name of the 
xlsx sheet to read in.}

\item{col_remove}{Optional argument. Vector of column number(s) to remove.}

\item{condition_cols}{Optional argument. Vector of column number(s) to
paste together to create the 'Conditon' column.}
}
\value{
Dataframe of phenotype information. A 'Condition' column of sample
    and treatment details concatenated together will be included if the 
    'condition_cols' argument is given.
}
\description{
Creates a dataframe containing phenotype information on the RPPA samples.
}
