% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_med.R
\name{norm_med}
\alias{norm_med}
\title{Median normalisation}
\usage{
norm_med(tidydf, col_name = "RFI")
}
\arguments{
\item{tidydf}{Tidy dataframe of RFI values. There should be NO technical
replicates.}

\item{col_name}{Column name of the median normalised values as a string.}
}
\value{
The same input dataframe but with a
}
\description{
Perform median normalisation on a tidy dataframe of RFI data.
}
\details{
This method assumes that the amount of all proteins measured in the 
experiment reflect the total protein amount of one sample. Thus, the 
median AB of a sample estimates total protein amount or 'sample loading'.

The normalisation calculation is: divide all raw values by the median value
of all AB signals of that sample is used to normalise the raw intensity
values for each AB. This may be biased when the number of ABs is <100.
}
