% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repConst.R
\name{plot_repConst}
\alias{plot_repConst}
\title{Replicate consistency}
\usage{
plot_repConst(tidydf, reps, title)
}
\arguments{
\item{tidydf}{Tidy dataframe of RFI data.}

\item{reps}{Vector of replicate names as string, to calculate consistency of.}

\item{title}{Title of the faceted barplot, as string.}

\item{filename}{Filename of the replicate consistency data (include .tsv at 
the end).}
}
\description{
Determine replicate consistency. by calculating the mean, standard deviation
and coefficient of variation (CV).
}
