% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{RFIperAB}
\alias{RFIperAB}
\alias{RFIperSample}
\title{Generate boxplots of RFI values}
\usage{
RFIperAB(tidydf, RFIcol = "RFI", fill_by)

RFIperSample(tidydf, RFIcol = "RFI", order, fill_by)
}
\arguments{
\item{tidydf}{Tidy dataframe of raw RFI values.}

\item{RFIcol}{Name of the column, as string, containing RFI values.}

\item{fill_by}{Optional argument. Name of the column, a string, to add to
the \code{fill} argument in ggplot's \code{aes}.}

\item{order}{Vector of sample names, in the order that they are to appear on
the graph.}
}
\value{
Creates plot in current graphics device.
}
\description{
Creates boxplots of raw RFI value for each antibody or each sample.
}
\section{Functions}{
\itemize{
\item \code{RFIperAB}: Creates a boxplot for each antibody.

\item \code{RFIperSample}: Creates a boxplot for each sample.
}}

