% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearmanPlot.R
\name{spearmanPlot}
\alias{spearmanPlot}
\title{Pairwise correlation coefficient plot}
\usage{
spearmanPlot(tidydf, RFI1, RFI2)
}
\arguments{
\item{tidydf}{Tidy dataframe with RFI values, antibody names, sample names
and any other optional columns.}

\item{RFI1}{Name of the column of the RFI values to rank, as string.}

\item{RFI2}{Optional argument. Name of the second column of RFI values to
rank, as string.}
}
\description{
Calculate the pairwise Spearman correlation between all possible combinations
of proteins and plot these (ordered) as a scatter plot.
}
\details{
This function performs the following steps:

\enumerate{
    \item For each protein, get the rank of it's expression level across
        all samples.
    \item Get matrix of all possible pairs of proteins.
    \item For each pair of proteins, calculate the Spearman correlation
        of their ranks.
    \item Plot all the (ordered) Spearman correlations for all the protein
        pairs as a scatter plot.
}

The distribution of the correlations can be indicative of the performance of
a normalisation technique. Differences in sample loading can result in 
artifically high correlations between protein pairs. You expect well 
normalised data to have similar positive and negative correlation 
coefficients. More information can be found in 
\href{http://journals.sagepub.com/doi/pdf/10.4137/CIN.S13329}{Liu et al.}.
}
