% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyData.R
\name{tidyData}
\alias{tidyData}
\title{Turn wide/matrix RFI data into tidy data}
\usage{
tidyData(df, ABnames, Batch = "A", ave_reps = FALSE, pheno)
}
\arguments{
\item{df}{Dataframe in the (matrix) format of the raw data.}

\item{ABnames}{Dataframe containing the antibody number - full
antibody name key-pairs.}

\item{Batch}{Batch code of the run. Can be single string or number. A column
of this code will be added to the output dataframe.}

\item{ave_reps}{Single logical indicating whether technical replicates 
(samples with the same names) should be averaged.}

\item{pheno}{Dataframe containing information on the sample phenotypes, to 
merge with the data using sample names.}
}
\value{
Tidy dataframe with columns containing full antibody names and 
phenotype information can also be merged into the base RFI dataframe.
}
\description{
Takes wide/matrix RFI data exported from Zeptosens where each row is an
sample and each column is an antibody and turns it into 
\href{https://cran.r-project.org/web/packages/tidyr/vignettes/tidy-data.html}{tidy data}.
}
\details{
A "Batch" column is added by default so that batches can be identified if
tidy dataframes from several runs are joined together.

If the exact same sample name appears multiple times, they represent 
technical replicates and can be averaged using 'ave_reps'.
}
