% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_plot.R
\name{time_plot}
\alias{time_plot}
\title{Time series graph}
\usage{
time_plot(tidydf, xcol, xlab, cond_col, log, ylab, scales)
}
\arguments{
\item{tidydf}{Tidy dataframe of RFI values, containing ONLY the 
conditions to plot.}

\item{xcol}{Column name of the desired x-axis variable, as string.}

\item{xlab}{x-axis label, as string.}

\item{cond_col}{Optional argument. Column name of the desired colour 
aesthetic, as string. If given will colour the lines by the condition.}

\item{log}{Logical indicating whether the RFI data should be logged 
(base 2). A small number (0.00001) is added to prevent taking the log
of zero.}

\item{ylab}{The y axis label as a single string.}
}
\description{
Creates faceted line graphs over time points for each antibody. Each graph
can also be coloured by a condition. One line graph will be produced per 
antibody and 12 graphs will be displayed per page. The graphs will be 
generated in the current graphics device.
}
