% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplots.R
\name{plotperSample}
\alias{plotperSample}
\alias{plotperAB}
\title{Barplots per sample/antibody}
\usage{
plotperSample(tidydf, RFIcol = "RFI", pdfoutput = TRUE)

plotperAB(tidydf, RFIcol = "RFI", pdfoutput = TRUE)
}
\arguments{
\item{tidydf}{Tidy dataframe of RFI values to plot.}

\item{RFIcol}{Column name of the RFI column as string.}

\item{pdfoutput}{Single logical indicating whether the plots should be
output to separate pdf file or printed in console.}
}
\description{
Creates a separate barplot of the RFI value for EACH sample or antibody and 
writes to either pdf called 'plotPerSample.pdf' or the current graphics
device.
}
\section{Functions}{
\itemize{
\item \code{plotperSample}: Generate a barplot of the RFI value for each 
antibody, one barplot per sample. Save in a file called 
'plotPerSample.pdf' if pdfoutput=TRUE.

\item \code{plotperAB}: Creates a barplot of the RFI value for each 
sample, one barplot per antibody. Writes to either a pdf called 
'plotPerAB.pdf' or the current graphics device.
}}

